requirement.txt
Flask==1.0.3
Flask_RESTful==0.3.7
Flask-Jsonpify==1.5.0
pandas==0.24.2
nltk==3.4.1
scikit-learn==0.21.0
mysql-connector==2.2.9